requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
distro==1.9.0
fastapi==0.115.5
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
openai==1.55.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
pypdfium2==4.30.0
python-dotenv==1.0.1
python-multipart==0.0.17
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.0
typing_extensions==4.12.2
uvicorn==0.32.1
